pyproject.toml
[tool.poetry]
name = "gitlablistpy"
version = "0.1.0"
description = ""
authors = ["Alexander Fast <git@alexanderfast.se>"]
readme = "README.md"
packages = [
    { include = "gitlablistpy" },
]

[tool.poetry.dependencies]
python = ">=3.9.6,<3.13"
python-gitlab = "^4.4.0"
pyinstaller = "^6.6.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
autopep8 = "^2.1.0"
autoflake8 = "^0.4.1"

[tool.poetry.scripts]
gitlablistpy = "gitlablistpy.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
